aanpassingen

+++ genres.php
@@ -33,8 +33,12 @@ if ($result = mysqli_query($conn, $sql)) {
                <!--<th>ID</th>-->
                <th>Genre naam</th>
                <th>Descriptie</th>
-               <th>Verwijder</th>
-               <th>Update</th>
+               <?php if (!empty($_SESSION)) : ?>
+                  <?php if ($_SESSION['role'] == "beheerder") : ?>
+                     <th>Verwijder</th>
+                     <th>Update</th>
+                  <?php endif ?>
+               <?php endif ?>
             </tr>
          </thead>
          <tbody>
@@ -43,13 +47,21 @@ if ($result = mysqli_query($conn, $sql)) {
                   <!--<td><?php echo $genre["genre_id"] ?></td>-->
                   <td><?php echo $genre["genreName"] ?></td>
                   <td><?php echo $genre["genreDescription"] ?></td>
-                  <td><a href="delete.php?genre_id=<?php echo $genre["genre_id"] ?>" class="shadow btn btn-danger shadow">Verwijder</a></td>
-                  <td><a href="update_gegevens.php?genre_id=<?php echo $genre["genre_id"] ?>" class="shadow btn btn-warning shadow">Update</a></td>
+                  <?php if (!empty($_SESSION)) : ?>
+                     <?php if ($_SESSION['role'] == "beheerder") : ?>
+                        <td><a href="delete.php?genre_id=<?php echo $genre["genre_id"] ?>" class="shadow btn btn-danger shadow">Verwijder</a></td>
+                        <td><a href="update_gegevens.php?genre_id=<?php echo $genre["genre_id"] ?>" class="shadow btn btn-warning shadow">Update</a></td>
+                     <?php endif ?>
+                  <?php endif ?>
                </tr>
             <?php endforeach; ?>
          </tbody>
       </table>
-      <a href="genre_toevoegen.php" class="w-100 btn btn-lg btn-warning shadow">Voeg een genre toe</a>
+      <?php if (!empty($_SESSION)) : ?>
+         <?php if ($_SESSION['role'] == "beheerder") : ?>
+            <a href="genre_toevoegen.php" class="w-100 btn btn-lg btn-warning shadow">Voeg een genre toe</a>
+         <?php endif ?>
+      <?php endif ?>
    </div>
 </body>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
